requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofile==3.9.0
    # via py-key-value-aio
aiofiles==24.1.0
    # via gradio
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   gradio
    #   httpx
    #   mcp
    #   openai
    #   py-key-value-aio
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==26.1.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.9
    # via fastmcp
backports-tarfile==1.2.0
    # via jaraco-context
beartype==0.22.9
    # via py-key-value-aio
brotli==1.2.0
    # via gradio
cachetools==7.0.5
    # via py-key-value-aio
caio==0.9.25
    # via aiofile
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.6
    # via
    #   authlib
    #   pyjwt
cyclopts==4.10.1
    # via fastmcp
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.4
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via
    #   anyio
    #   fastmcp
fastapi==0.135.2
    # via
    #   gradio
    #   openenv-core
fastmcp==3.1.1
    # via openenv-core
ffmpy==1.0.0
    # via gradio
filelock==3.25.2
    # via huggingface-hub
fsspec==2026.3.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==6.10.0
    # via openenv-core
gradio-client==2.4.0
    # via
    #   gradio
    #   hf-gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-gradio==0.3.0
    # via gradio
hf-xet==1.4.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   mcp
    #   openai
    #   openenv-core
    #   safehttpx
httpx-sse==0.4.3
    # via mcp
huggingface-hub==1.8.0
    # via
    #   gradio
    #   gradio-client
    #   openenv-core
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via
    #   keyring
    #   opentelemetry-api
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.1.2
    # via keyring
jaraco-functools==4.4.0
    # via keyring
jinja2==3.1.6
    # via gradio
jiter==0.13.0
    # via openai
jsonref==1.1.0
    # via fastmcp
jsonschema==4.26.0
    # via mcp
jsonschema-path==0.4.5
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.7.0
    # via py-key-value-aio
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
mcp==1.26.0
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via
    #   jaraco-classes
    #   jaraco-functools
numpy==2.2.6
    # via
    #   gradio
    #   pandas
openai==2.30.0
    # via openenv-core
openapi-pydantic==0.5.1
    # via fastmcp
openenv-core==0.2.3
    # via openenv-my-env (pyproject.toml)
opentelemetry-api==1.40.0
    # via fastmcp
orjson==3.11.7
    # via gradio
packaging==26.0
    # via
    #   fastmcp
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==2.3.3
    # via gradio
pathable==0.5.0
    # via jsonschema-path
pillow==12.1.1
    # via gradio
platformdirs==4.9.4
    # via fastmcp
py-key-value-aio==0.4.4
    # via fastmcp
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   fastmcp
    #   gradio
    #   mcp
    #   openai
    #   openapi-pydantic
    #   openenv-core
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pydub==0.25.1
    # via gradio
pygments==2.20.0
    # via rich
pyjwt==2.12.1
    # via mcp
pyperclip==1.11.0
    # via fastmcp
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.22
    # via
    #   gradio
    #   mcp
pytz==2026.1.post1
    # via
    #   gradio
    #   pandas
pywin32==311
    # via mcp
pywin32-ctypes==0.2.3
    # via keyring
pyyaml==6.0.3
    # via
    #   fastmcp
    #   gradio
    #   huggingface-hub
    #   jsonschema-path
    #   openenv-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.33.0
    # via openenv-core
rich==14.3.3
    # via
    #   cyclopts
    #   fastmcp
    #   openenv-core
    #   rich-rst
    #   typer
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safehttpx==0.1.7
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
sse-starlette==3.3.4
    # via mcp
starlette==0.52.1
    # via
    #   fastapi
    #   gradio
    #   mcp
    #   sse-starlette
tomli==2.4.1
    # via
    #   cyclopts
    #   openenv-core
tomli-w==1.2.0
    # via openenv-core
tomlkit==0.13.3
    # via gradio
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   openai
typer==0.24.1
    # via
    #   gradio
    #   hf-gradio
    #   huggingface-hub
    #   openenv-core
typing-extensions==4.15.0
    # via
    #   anyio
    #   cryptography
    #   cyclopts
    #   exceptiongroup
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   mcp
    #   openai
    #   opentelemetry-api
    #   py-key-value-aio
    #   pydantic
    #   pydantic-core
    #   pyjwt
    #   referencing
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   fastapi
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
uncalled-for==0.2.0
    # via fastmcp
urllib3==2.6.3
    # via requests
uvicorn==0.42.0
    # via
    #   fastmcp
    #   gradio
    #   mcp
    #   openenv-core
watchfiles==1.1.1
    # via fastmcp
websockets==16.0
    # via
    #   fastmcp
    #   openenv-core
zipp==3.23.0
    # via importlib-metadata
